Updated readme and pecl pathway
Ran into two problems when I tried installing on Ubuntu 15.10 box.  The first issue was that composer kept giving me an error telling me that I must use http-secure, so the file paths are updated to use https instead of http for pecl.

Sadly the windows.php.net does not have https support, or not any support that I can connect to.

The second is that it's not documented that you might need to run the command as sudo.

Ran tests and they passed.

Signed-off-by: Michael Soileau <webart.video@gmail.com>