Properly detect compiler on Windows
At least for the binary PECL package downloads, the compiler on Windows
has to be something like "vc15" or "vs16".  However, the PHP info
reports the compiler as something like "Visual C++ 2017".  Instead of
mapping this information back to what we need, we use the "PHP
Extension Build" info, which already contains the desired compiler
name.